requirements.txt
# Core
pytz==2019.*
Django==2.1.*
django-filter==2.1.*
django-configurations==2.*

djangorestframework==3.9.*
drf-yasg==1.*
django-cors-headers==2.*

django-stubs==0.*
djangorestframework-stubs==0.*
gunicorn==19.9.*

# For the persistence stores
psycopg2-binary==2.*
dj-database-url==0.5.*

# Model Tools
django-model-utils==3.*
django-simple-history==2.7.*

# Developer Tools
ipdb==0.11
ipython==7.2.0

# logging
sentry-sdk==0.*

# tasks
celery[redis]==4.3.*
django-celery-beat==1.3.*
flower==0.9.*

# cache
django-redis==4.10.*
django-cachalot==2.1.*
